Refactoring settings module
- Move from file to folder structure
- Move json and sqlite to separate sub-modules
